Plumb resolved_url through params, simplify rewrite check, consolidate tests
